requirements.txt
alembic==1.0.8
alpha-vantage==2.1.0
amqp==2.4.2
asn1crypto==0.24.0
awsebcli==3.15.0
backports.ssl-match-hostname==3.7.0.1
billiard==3.5.0.5
blessed==1.15.0
blinker==1.4
botocore==1.12.134
cached-property==1.5.1
celery==4.2.1
cement==2.8.2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.6.1
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
functools32==3.2.3.post2
future==0.16.0
idna==2.5
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
kombu==4.4.0
Mako==1.0.7
MarkupSafe==1.1.1
newsapi==0.1.1
newsapi-python==0.2.3
numpy==1.16.2
pandas==0.24.2
pathspec==0.5.9
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
PyYAML==3.13
redis==3.2.1
requests==2.20.1
rq==0.13.0
semantic-version==2.5.0
simplifycommerce-sdk-python==1.6.0
six==1.11.0
SQLAlchemy==1.3.1
termcolor==1.1.0
texttable==0.9.1
twilio==6.26.0
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.14.1
WTForms==2.2.1
